requirements.txt
streamlit==1.39.0
pillow==10.4.0
reportlab==4.2.2
numpy==1.26.4
